# AndNode.substitute short circuit