@roelofvandijk 2023-09-06
How about this?

```
def substitute(self, var_vals: Dict[Variable, Node]) -> Node:
  subbed = ()
  for node in self.nodes:
    if not (sub:=node.substitute(var_vals)): return NumNode(0)
    subbed += (sub,)
  return Variable.ands(subbed)
```